more robuts delete

+++ Tests/CatalogItems/Rest/Out-RsRestCatalogItem.Tests.ps1
@@ -44,9 +44,9 @@ Describe "Out-RsRestCatalogItem" {
 
     AfterEach {
         Remove-RsCatalogItem -ReportServerUri $reportServerUri -RsFolder $rsFolderPath -Confirm:$false
-        if ($localFolderPath -ne "")
+        if ($localFolderPath -ne $null -and $localFolderPath .Length -ne 0 -and (Test-Path $localFolderPath))
         {
-            Remove-Item -Path $localFolderPath -Recurse
+            Remove-Item -Path $localFolderPath  -Recurse
         }
     }
